Notes can be created and moved around.
ToDo type has a zIndex, color, and colorSecondary property.

When the user generates a new note, the note will be a random color.

When the user picks up the note via the top area of the note, they can move it around and it will have the highest zIndex on the container.